chore: move app to index and extract notifications
