--- src/components/App/styles.ts
@@ -1,5 +1,0 @@
-import styled from 'styled-components';
-
-export const Container = styled.div`
-  padding: 8px;
-`;
