--- src/components/NotificationArticles/index.tsx
@@ -1,244 +1,0 @@
-import React, { useEffect, useRef } from 'react';
-import { audioPlayer } from '../../audio-player';
-import { ButtonInfo } from '../../lib/types';
-import { Container } from './styles';
-
-type Props = {
-  buttonInfoList: ButtonInfo[];
-};
-
-export function NotificationArticles({ buttonInfoList }: Props) {
-  const ref = useRef<HTMLDivElement | null>(null);
-
-  useEffect(() => {
-    if (!buttonInfoList || !buttonInfoList.length) {
-      return;
-    }
-
-    const notificationRowEls = ref.current!;
-    const buttons = notificationRowEls.querySelectorAll<HTMLButtonElement>('.sounds');
-
-    for (const button of buttons) {
-      button.onclick = () => {
-        const fileName = button.getAttribute('data-file');
-        const audioId = buttonInfoList.findIndex((buttonInfo) => buttonInfo['file-name'] === fileName);
-
-        if (audioId === undefined) {
-          console.error(`${fileName} play failed`);
-
-          return;
-        }
-
-        audioPlayer.emitAudioId(audioId);
-      };
-    }
-  }, [buttonInfoList]);
-
-  return (
-    <Container ref={ref}>
-      <span className="important-notify">
-        <b>名取さな単独イベント『さなのばくたん。』3月7日</b>
-      </span>
-      <br />
-      中止になりました。
-      <a href="https://bakutan.natorisana.com/cancellation.html" target="_blank" rel="noopener noreferrer">
-        イベント開催中止のお知らせ
-      </a>
-      <br />→{' '}
-      <a href="https://www.youtube.com/watch?v=M5zxb_3d2FA">
-        『なのばくさん。～イベント中止になっちゃったので新3Dお披露目しまスペシャル～』20時から
-      </a>
-      <hr style={{ margin: '1em 0' }} />
-      <span className="important-notify">
-        <b>名取さなオンリー同人誌即売会『名取さマ～？』2020/3/7開催！</b>
-      </span>
-      <br />
-      <button type="button" className="sounds" data-file="漢字テスト：おまいら自重[　　　]汁ｗｗｗｗｗ/何とは言いませんが">
-        何とは言いませんが
-      </button>
-      <button type="button" className="sounds" data-file="漢字テスト：おまいら自重[　　　]汁ｗｗｗｗｗ/噂を聞きつける">
-        噂を聞きつける
-      </button>
-      <p>
-        <a href="https://twitter.com/natorisummer">【イベント情報はこちら】</a>
-        <a href="https://natorisummer.site/circles/">【参加サークル一覧はこちら】</a>
-      </p>
-      <hr style={{ margin: '1em 0' }} />
-      <span className="important-notify">
-        <b>『さなのおうた。』楽曲配信開始！</b>
-      </span>
-      <br />
-      <button type="button" className="sounds" data-file="性の6時間は名取と子作りしようよっ！2019/さなのおうた。配信開始！">
-        さなのおうた。配信開始！
-      </button>
-      <button type="button" className="sounds" data-file="性の6時間は名取と子作りしようよっ！2019/Spotifyもある">
-        Spotifyもある
-      </button>
-      <p>
-        <a href="https://music.apple.com/jp/album/%E3%81%95%E3%81%AA%E3%81%AE%E3%81%8A%E3%81%86%E3%81%9F-single/1490088141">
-          Apple Music/iTunesはこちら
-        </a>
-        <a href="https://open.spotify.com/album/2hRbK2th3yOuwrKSUV2LTv">Spotifyはこちら</a>
-        <a href="https://play.google.com/store/music/album?id=Bixi2p3tciazqbpzwtzw5sfmpqe">Google Playはこちら</a>
-      </p>
-      <p>
-        <a href="https://linktr.ee/natori_sana">DLリンクまとめはこちら</a>
-        <a href="https://www.youtube.com/watch?v=UADoH5n1RLM">さなのおうた。(YouTube)</a>
-      </p>
-      <hr style={{ margin: '1em 0' }} />
-      <span className="important-notify">
-        <b>『#名取さな ちゃん1/7スケールフィギュア』予約受付中！！</b>
-      </span>
-      <br />
-      <button type="button" className="sounds" data-file="どうも、知的で可愛い名取さなです。/名取さなフィギュア化決定！">
-        名取さなフィギュア化決定！
-      </button>
-      <button type="button" className="sounds" data-file="どうも、知的で可愛い名取さなです。/（下から見れますよ）">
-        （下から見れますよ）
-      </button>
-      <button type="button" className="sounds" data-file="どうも、知的で可愛い名取さなです。/（下から見れます）">
-        （下から見れます）
-      </button>
-      <button type="button" className="sounds" data-file="どうも、知的で可愛い名取さなです。/続報を待て">
-        続報を待て
-      </button>
-      <button type="button" className="sounds" data-file="どうも、知的で可愛い名取さなです。/しばし待て">
-        しばし待て
-      </button>
-      <button type="button" className="sounds" data-file="台風くん！後生だから/販売されます">
-        販売されます
-      </button>
-      <button type="button" className="sounds" data-file="台風くん！後生だから/フィギュアそろそろらしい">
-        フィギュアそろそろらしい
-      </button>
-      <button type="button" className="sounds" data-file="台風くん！後生だから/生産を遅らせるアヘエビ">
-        生産を遅らせるアヘエビ
-      </button>
-      <button type="button" className="sounds" data-file="台風くん！後生だから/アヘエビを持たせたかった">
-        アヘエビを持たせたかった
-      </button>
-      <button type="button" className="sounds" data-file="台風くん！後生だから/量産だとアヘエビ持てない">
-        量産だとアヘエビ持てない
-      </button>
-      <p>
-        <a href="https://www.amiami.jp/top/detail/detail?gcode=FIGURE-054527">あみあみはこちら（特典はアクリルストラップ）</a>
-        <a href="https://www.gamers.co.jp/pn/pd/10476310/">ゲーマーズはこちら（特典はA3タペストリー）</a>
-        <a href="https://www.suruga-ya.jp/product/detail/602185083">駿河屋はこちら（特典はラバーキーホルダー）</a>
-      </p>
-      <p>
-        <a href="https://twitter.com/sana_natori/status/1197504578431832064">特典情報はこちら</a>
-        <a href="https://twitter.com/wanderer1127">わんだらー様のTwitterはこちら</a>
-      </p>
-      <hr style={{ margin: '1em 0' }} />
-      <span className="important-notify">
-        <b>ば〜ちゃるな〜す名取さなLINEスタンプ第1弾販売中！</b>
-      </span>
-      <br />
-      <p>
-        <a href="https://t.co/6zv09p2Kgn">購入はこちら</a>（※LINE スタンプはしゃべりません）
-      </p>
-      <span>※ボタンを押して再生はされますが、仮置きです。近い内に変更すると思います🙏</span>
-      <br />
-      <button type="button" className="sounds" data-file="おえかきカキカキでワロタｗ/おはようございナース">
-        おはようございナース
-      </button>
-      <button type="button" className="sounds" data-file="破壊のち創造ってワケでしてｗ/ニマァ">
-        二マァ
-      </button>
-      <button type="button" className="sounds" data-file="3D名取のﾌﾆｬﾗｶｻｷﾞｮｰﾊｲｼﾝ/やっていくわよ">
-        やっていくわよ
-      </button>
-      <button type="button" className="sounds" data-file="今日は9月31日/えーんえんえんおいおいおい">
-        えーんえんえんおいおいおい
-      </button>
-      <button type="button" className="sounds" data-file="バイノーラルテスト配信/押し掛け">
-        メーターまわってますよ
-      </button>
-      <button type="button" className="sounds" data-file="Let‘s GO！名取さな＃1/完全に理解した">
-        完全に理解した
-      </button>
-      <button type="button" className="sounds" data-file="首の長さを修正したことを懺悔してみた/何の話してたっけ">
-        何の話してたっけ
-      </button>
-      <button type="button" className="sounds" data-file="すこ/すこ04">
-        すこすこのすこ
-      </button>
-      <br />
-      <button type="button" className="sounds" data-file="3D名取のﾌﾆｬﾗｶｻｷﾞｮｰﾊｲｼﾝ/なで1">
-        何で（なで）
-      </button>
-      <button type="button" className="sounds" data-file="ハロウィンイベ蜜チューチューでワロタ/やっぴ～！">
-        やっぴ～！
-      </button>
-      <button type="button" className="sounds" data-file="せんせえとないしょばなし/正解正解">
-        正解正解正解
-      </button>
-      <button type="button" className="sounds" data-file="ネコのさばきは突然にくる/僕もそう思いますにゃ">
-        僕もそう思いますにゃ
-      </button>
-      <button type="button" className="sounds" data-file="3D名取のﾎﾆｬﾎﾆｬﾄﾞﾛｰｲﾝｸﾞ/呼吸頑張ってる">
-        呼吸頑張ってる
-      </button>
-      <button type="button" className="sounds" data-file="れいや～ずおぶふぃあ～Part4/了解！やります！">
-        了解
-      </button>
-      <button type="button" className="sounds" data-file="破壊のち創造ってワケでしてｗ/売っちゃった2">
-        売っちゃった
-      </button>
-      <button type="button" className="sounds" data-file="Let‘s GO！限界オタク＃2/いいね！2">
-        いいね！
-      </button>
-      <button type="button" className="sounds" data-file="４まんにんってなんまんにん？/偉すぎぃ！">
-        偉すぎぃ！
-      </button>
-      <br />
-      <button type="button" className="sounds" data-file="めうから奪ったはんこでフェスやるめう。/キレちゃったね～">
-        キレちゃったね～
-      </button>
-      <button type="button" className="sounds" data-file="破壊のち創造ってワケでしてｗ/オエってしてくる">
-        オエってしてくる
-      </button>
-      <button type="button" className="sounds" data-file="6万人記念でホラゲ/息切れナース02">
-        ハァハァ
-      </button>
-      <button type="button" className="sounds" data-file="つよつよPCチャレンジゲリラ/大変感謝しております">
-        感謝
-      </button>
-      <button type="button" className="sounds" data-file="破壊のち創造ってワケでしてｗ/賢くなったネ">
-        賢くなったネ
-      </button>
-      <button type="button" className="sounds" data-file="毎年カップル板クリスマスコピペを見てブチ切れていた女の配信/メリークリスマス2">
-        クリスマス
-      </button>
-      <button type="button" className="sounds" data-file="6万人記念でホラゲ/マ？02">
-        マァ〜？
-      </button>
-      <button type="button" className="sounds" data-file="登録者2万人達成で泣く名取/マ！？">
-        マ！？
-      </button>
-      <br />
-      <button type="button" className="sounds" data-file="破壊のち創造ってワケでしてｗ/すまねぇ〜">
-        すまねぇ〜
-      </button>
-      <button type="button" className="sounds" data-file="LINEすたんぷをつくるふぁむふぁたる/はい・・・">
-        はい・・・
-      </button>
-      <button type="button" className="sounds" data-file="てね/てね03">
-        ってね
-      </button>
-      <button type="button" className="sounds" data-file="平成最後の配信だが礼はいらんゾｗってね/フリーズしたハムスター">
-        フリーズしたハムスター
-      </button>
-      <button type="button" className="sounds" data-file="Let‘s GO！メスガキ四天王序列三位/もう許さねえからな">
-        許さねえ
-      </button>
-      <button type="button" className="sounds" data-file="Let‘s GO！名取さな＃1/笑い声9">
-        笑い声
-      </button>
-      <button type="button" className="sounds" data-file="ヒーラー職だからです/おつかれさな〜">
-        おつかれさな〜
-      </button>
-      <hr style={{ margin: '1em 0' }} />
-    </Container>
-  );
-}
