+++ src/components/Notifications/styles.ts
