+++ src/components/index.ts
@@ -1,11 +1,10 @@
-export * from './App';
 export * from './UpdateLog';
 export { FixedHeader } from './FixedHeader';
 export { Header } from './Header';
-export * from './NotificationArticles';
 export * from './PostArticles';
 export * from './Broadcasts';
 export * from './BroadCastLinkList';
 export * from './AudioMenu';
 export * from './NatoriSana';
 export * from './Button';
+export * from './Notifications';
